modules/06-big-data-architecture/03-implementations/mini-platform/services/worker/requirements.txt
kafka-python==2.0.2
psycopg2-binary==2.9.9
minio==7.2.5
clickhouse-driver==0.2.7
